% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bluertopo-package}
\alias{bluertopo-package}
\title{bluertopo: Download and extract NOAA BlueTopo bathymetry with terra}
\description{
\code{bluertopo} discovers NOAA BlueTopo source tiles, downloads original
GeoTIFF/RAT assets with checksum validation, and opens selected raster bands
as file-backed \code{terra} objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/el-cordero/bluer-topo}
  \item Report bugs at \url{https://github.com/el-cordero/bluer-topo/issues}
}

}
\author{
\strong{Maintainer}: Eliseo Custodio \email{eliseo@example.com}

}
\keyword{internal}
